refactor: consolidate label flip + notify into single helper
Move both the label flip and the notify dispatch into flipLabelAndNotify so single-story-close.js stays leaner.
