+++ .agents/scripts/lib/single-story/story-merged-notify.js
@@ -1,15 +1,60 @@
 /**
- * story-merged-notify.js — best-effort `story-merged` dispatch for the
- * standalone-Story close path. Mirrors the event name used by
- * `post-merge-pipeline.notificationPhase` so operator subscriptions cover
- * both Epic-attached and standalone Stories.
+ * story-merged-notify.js — flip Story label to agent::done and fire the
+ * `story-merged` event for a standalone-Story close. Mirrors the event
+ * name used by `post-merge-pipeline.notificationPhase` so operator
+ * subscriptions cover both Epic-attached and standalone Stories.
+ *
+ * Both the label flip and the notify dispatch are best-effort: failures
+ * are logged and swallowed so neither a flaky GitHub API nor a flaky
+ * webhook ever fails the close.
  */
 
 import { Logger } from '../Logger.js';
+import { notify as defaultNotify } from '../../notify.js';
 
-export async function dispatchStoryMergedNotify({
+export async function flipLabelAndNotify({
+  provider,
   notifyFn,
-  labelFlipped,
+  storyId,
+  story,
+  prUrl,
+  autoMergeEnabled,
+  autoMergeReason,
+  orchestration,
+  progress,
+}) {
+  const labelFlipped = await flipLabel(provider, storyId, story, progress);
+  if (!labelFlipped) return;
+  await fireStoryMergedNotify({
+    notifyFn: notifyFn ?? defaultNotify,
+    storyId,
+    story,
+    prUrl,
+    autoMergeEnabled,
+    autoMergeReason,
+    orchestration,
+    provider,
+  });
+}
+
+async function flipLabel(provider, storyId, story, progress) {
+  try {
+    const labels = (story.labels || [])
+      .filter((l) => !l.startsWith('agent::'))
+      .concat('agent::done');
+    await provider.updateTicket(storyId, { labels });
+    progress?.('LABELS', `🏷️  Story #${storyId} → agent::done`);
+    return true;
+  } catch (err) {
+    Logger.error(
+      `[single-story-close] ⚠️ Failed to flip Story labels: ${err?.message ?? err}`,
+    );
+    return false;
+  }
+}
+
+async function fireStoryMergedNotify({
+  notifyFn,
   storyId,
   story,
   prUrl,
@@ -18,7 +63,6 @@ export async function dispatchStoryMergedNotify({
   orchestration,
   provider,
 }) {
-  if (!labelFlipped) return;
   const autoMergeNote = autoMergeEnabled
     ? 'auto-merge enabled — GitHub will squash-merge when required checks pass'
     : `auto-merge not enabled (${autoMergeReason ?? 'unknown'}) — operator merges via GitHub UI`;
